requirements.txt
asgiref==3.7.2
cffi==1.16.0
charset-normalizer==3.3.2  
cryptography==41.0.7       
distro==1.9.0
Django==5.0
django-restframework==0.0.1
djangorestframework==3.14.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
pandas==2.1.4
pdfminer.six==20231228     
pycparser==2.21
PyPDF2==3.0.1
pyproject==1.3.1
python-dateutil==2.8.2
pytz==2023.3.post1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tabula-py==2.9.0
tzdata==2023.4